completed

+++ lib/model/data_model/batch_fee_invoice_list.dart
@@ -116,7 +116,7 @@ class Datum {
   int? monthClassesConductedCount;
   int? monthAttendancePresentCount;
   int? cycleAttendancePresentCount;
-  int? pendingAmount;
+  double? pendingAmount;
   dynamic totalNoOfClasses;
   By? writtenOffBy;
   StudentDetail? studentDetail;
